Introduce Analytical Benchmarking and Metrics
We do not have functionality to control and run concurrent reads
functionality within Lucene today, and with the ongoing effort
to extend that functionality, it is important to have a generic
benchmark which enabled this benchmarking. Also, we do not calculate
P50, P90 and P100 metrics which are useful for analytical queries.
This commit introduces these functionalities

Signed-off-by: Atri Sharma <atris@amazon.com>