Add more interesting tasks for top-k hits on boolean queries.
Looking into the [Tantivy benchmark](https://tantivy-search.github.io/bench/)
highlighted that there are interesting cases that are not covered by our
nightly benchmarks. For instance, MAXSCORE splits clauses into essential and
non-essential clauses, but our nightly benchmarks only really benchmark the
case when there is a single essential clause (with `OrHighMed` because there is
one clause whose scores dominates, and with `OrHighHigh` because there are so
many documents that have both terms that `MAXSCORE` quickly realizes that the
clause with the maximum score is required for a hit to be competitive anyway).

This adds a few more categories whose performance would be interesting to
track. I'm expecting the performance of these queries to often react very
differently to changes in query evaluation compared to our historic
`OrHighHigh`, `OrHighMed`, `AndHighHigh` and `AndHighMed` queries.