Add Cohere/wikipedia-22-12-en-embeddings dataset for vector search
This dataset contains vectors of higher dimenstions of 768.
This dataset contains a pre-processed version from Wikipedia suitable for semantic search.

The problem with current datasets generated by infer_token_vectors.py file is that
they are generated for single tokens out of context. Thus, they don't represent the true
embeddigns, and hence the performance of vector search on them is not good.

 Cohere/wikipedia-22-12-en-embeddings dataset represent true embeddings and we can
show very good performance of vector search on them.