Fix bias towards competitor in searchBench
This makes competitor and baseline run in alternating order
instead of keeping the same order for all iterations.
This removes non-trivial bias in tests that run the system at
full load. The bias comes from the fact that generating the report
after each iteration takes macroscopic time but only utilizes a single
CPU cor. This allows the CPU to cool down giving the competitor that
runs first an advantage on a fully loaded system because it starts on
a cooler CPU. There's also other potential biases like that from
page-cache utilization and the like that may or may not advantage the
first or second run.
I could clearly reproduce a bias for the competitor before this change,
making things run in alterating order removes this effect in my testing.