Pollute call sites before running benchmarks.
Because nightly benchmarks only test a small set of scenarios, the JVM may end
up over-optimizing query evaluation. For instance, it only runs with
BM25Similarity, sorting tasks only run against a TermQuery, filtered vector
search only exercises the approximate path, not the exact path, etc.

This tries to make the benchmark more realistic by running some cheap queries
before running bencharks, whose goal is to pollute call sites so that they are
not all magically monomorphic.

This will translate in a drop in performance for some tasks, but hopefully we
can recover some of it in the future.

Related PR:
 - https://github.com/apache/lucene/pull/14968 where we suspected the speedup
   to be due to specialization making a call site monomorphic in nightly
   benchmarks that would not be monomorphic in the real world,
 - https://github.com/apache/lucene/pull/15039 where we are trying to improve
   behavior with several different similarity impls but the benchmarks only
   show a small improvement since they always run with BM25Similarity.